packages/stable-coin-xrp/token-backend/requirements.txt
Flask==3.0.3
xrpl-py==2.6.0
